Add fix for removing and adding rows.
Previously, logic depended on lastChild and firstChild, for some reason, append and prepend methods
did not add elemnt in a correct sequence. At least, result (positions of elements in dom) was unpredictable.
Instead of relying on position of element in DOM, 'rows' of elements beign stored in map by [rowIndex]:rows.